+++ DragaliaAPI.Photon.Shared/Models/HeroParamResponse.cs
@@ -8,7 +8,7 @@ namespace DragaliaAPI.Photon.Shared.Models
     {
         public int ActorNr { get; set; }
 
-        public int ViewerId { get; set; }
+        public long ViewerId { get; set; }
 
         public List<IEnumerable<HeroParam>> HeroParamLists { get; set; } =
             new List<IEnumerable<HeroParam>>();
